==> 1_Front_end/03_Concat_with_label+Pruning_BatchTest_.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn import preprocessing
import time

### Loading Data and Processing

#############Loading raw data#############
df = pd.read_csv("data/flow_s_alltime_Dropna.csv")
df.shape[0]

#############Loading Label#############
dflabel_blstm = pd.read_csv("Results/ex_label&model-blstm.csv")
dflabel_blstm_cnn = pd.read_csv("Results/ex_label&model-blstm-cnn.csv")
dflabel_cnn = pd.read_csv("Results/ex_label&model-cnn.csv")
dflabel_cnn_blstm = pd.read_csv("Results/ex_label&model-cnn-blstm.csv")

dflabel_tfxgb = pd.read_csv("Results/ex_label_tf_xgb.csv")
dflabel_tfrf = pd.read_csv("Results/ex_label_tf_rf.csv")
dflabel_tflsvm = pd.read_csv("Results/ex_label_tf_lsvm.csv")

dflabel_wvxgb = pd.read_csv("Results/ex_label_wv_xgb.csv")
dflabel_wvrf = pd.read_csv("Results/ex_label_wv_rf.csv")
dflabel_wvlsvm = pd.read_csv("Results/ex_label_wv_lsvm.csv")

dflabel_blstm.shape[0]

#############Concat#############
dfconcat_blstm = pd.concat([df, dflabel_blstm], axis=1)
dfconcat_blstm_cnn = pd.concat([df, dflabel_blstm_cnn], axis=1)
dfconcat_cnn = pd.concat([df, dflabel_cnn], axis=1)
dfconcat_cnn_blstm = pd.concat([df, dflabel_cnn_blstm], axis=1)

dfconcat_tfxgb = pd.concat([df, dflabel_tfxgb['label1']], axis=1)
dfconcat_tfrf = pd.concat([df, dflabel_tfrf['label1']], axis=1)
dfconcat_tflsvm = pd.concat([df, dflabel_tflsvm['label1']], axis=1)

dfconcat_tfxgb.rename(columns={'label1':'label'},inplace=True) 
dfconcat_tfrf.rename(columns={'label1':'label'},inplace=True) 
dfconcat_tflsvm.rename(columns={'label1':'label'},inplace=True) 

dfconcat_wvxgb = pd.concat([df, dflabel_wvxgb['label1']], axis=1)
dfconcat_wvrf = pd.concat([df, dflabel_wvrf['label1']], axis=1)
dfconcat_wvlsvm = pd.concat([df, dflabel_wvlsvm['label1']], axis=1)

dfconcat_wvxgb.rename(columns={'label1':'label'},inplace=True) 
dfconcat_wvrf.rename(columns={'label1':'label'},inplace=True) 
dfconcat_wvlsvm.rename(columns={'label1':'label'},inplace=True) 

#############Importance! sort_values and reset_index#############
dfconcat_blstm = dfconcat_blstm.sort_values(["label"],ascending=True)
dfconcat_blstm_cnn = dfconcat_blstm_cnn.sort_values(["label"],ascending=True)
dfconcat_cnn = dfconcat_cnn.sort_values(["label"],ascending=True)
dfconcat_cnn_blstm = dfconcat_cnn_blstm.sort_values(["label"],ascending=True)
dfconcat_tfxgb = dfconcat_tfxgb.sort_values(["label"],ascending=True)###
dfconcat_tfrf = dfconcat_tfrf.sort_values(["label"],ascending=True)###
dfconcat_tflsvm = dfconcat_tflsvm.sort_values(["label"],ascending=True)###
dfconcat_wvxgb = dfconcat_wvxgb.sort_values(["label"],ascending=True)###
dfconcat_wvrf = dfconcat_wvrf.sort_values(["label"],ascending=True)###
dfconcat_wvlsvm = dfconcat_wvlsvm.sort_values(["label"],ascending=True)###

dfconcat_blstm.reset_index(drop=True, inplace=True)
dfconcat_blstm_cnn.reset_index(drop=True, inplace=True)
dfconcat_cnn.reset_index(drop=True, inplace=True)
dfconcat_cnn_blstm.reset_index(drop=True, inplace=True)
dfconcat_tfxgb.reset_index(drop=True, inplace=True)
dfconcat_tfrf.reset_index(drop=True, inplace=True)
dfconcat_tflsvm.reset_index(drop=True, inplace=True)
dfconcat_wvxgb.reset_index(drop=True, inplace=True)
dfconcat_wvrf.reset_index(drop=True, inplace=True)
dfconcat_wvlsvm.reset_index(drop=True, inplace=True)


dfconcat_wvlsvm.head()
dfconcat_wvrf.head()
dfconcat_wvxgb.head()
dfconcat_tflsvm.head()
dfconcat_tfrf.head()
dfconcat_tfxgb.head()
dfconcat_blstm_cnn.head()

### Defining Parameters and Functions

#############Defining Parameters and Functions#############
dfabcorrect=['10.49.223.194','10.49.231.206','10.49.212.126','123.58.173.111', '10.49.212.162','10.49.21.15',
             '10.49.253.233','10.49.253.118','10.49.253.147','10.49.253.204','10.49.253.210','10.49.253.225',
             '10.49.253.252','10.49.253.35','10.49.253.86','10.49.253.87','10.56.144.126','10.49.253.175',
             '10.56.34.63','10.49.253.157','10.56.148.238','10.56.4.86','10.56.68.148','10.56.115.19',
             '10.24.139.131','10.49.126.199']

################Change here to test the results of data leakage incidents identification by web front-end anomaly detection only
# dfabcorrect=['10.49.253.233','10.49.253.35','10.49.223.194','10.49.212.162','10.49.21.15',
#             '10.49.141.206','123.58.173.111']


dfabcorrect_set = set(dfabcorrect)

#######Normalization Function
min_max_scaler_01 = preprocessing.MinMaxScaler(feature_range=(0,1), copy=True)

#######Pruning Function
def Pruning(modelpred_,DelNum):
    if DelNum == 0:
        return modelpred_
    
    modelpred__For_Pruning = modelpred_[modelpred_['label']<0.5][['time','source_ip','destination_ip','num']]
    
    ab_finalsort_half = modelpred__For_Pruning
    ################Important  Converts to strings to match records
    ab_finalsort_half['source_ip'] = ab_finalsort_half['source_ip'].apply(str)
    ab_finalsort_half['destination_ip'] = ab_finalsort_half['destination_ip'].apply(str)
    ab_finalsort_half = ab_finalsort_half[ab_finalsort_half['destination_ip'].str.startswith('10.')]
    
    column_uniques_sip = ab_finalsort_half['source_ip'].unique()
    groups_sip = ab_finalsort_half.groupby('source_ip')

    dfkeep = pd.DataFrame(columns=('sip','tihNum','dipNum','numSum'))

    for column_sip in column_uniques_sip:
        dfexpi_sip = groups_sip.get_group(column_sip)
        dfexpi_sip.reset_index(drop=True, inplace=True)
        dfkeep = dfkeep.append({'sip':column_sip,'tihNum':dfexpi_sip['time'].nunique(),'dipNum':dfexpi_sip['destination_ip'].nunique(),'numSum':dfexpi_sip['num'].sum()}, ignore_index=True)

    mmfeature=[x for x in dfkeep.columns if x not in ['sip']]
    for col in dfkeep[mmfeature]:
        dfkeep[[col]]=min_max_scaler_01.fit_transform(dfkeep[[col]])
    
    ###Remove sips that do not meet multi-tih and multi-dip
    dfkeep = dfkeep.drop(index = dfkeep[dfkeep['dipNum'] == 0].index.tolist())
    dfkeep = dfkeep.drop(index = dfkeep[dfkeep['tihNum'] == 0].index.tolist())
    
    dfkeep['nor_score'] = dfkeep['tihNum'] + dfkeep['dipNum'] + dfkeep['numSum']
    dfkeep.sort_values(by="nor_score" , inplace=True, ascending=False)
    dfkeep.reset_index(drop=True, inplace=True)

    dfkeep["nor_score"] = pd.to_numeric(dfkeep["nor_score"],errors='coerce')

    dfdrop_ip = dfkeep[0:DelNum]['sip']
    dfdrop_iplist = dfdrop_ip.tolist()
    
    for ip in dfdrop_iplist:
        modelpred_ = modelpred_.drop(index = modelpred_[modelpred_['source_ip'] == ip].index.tolist())###Delete Scanner-related records
    modelpred_.reset_index(drop=True, inplace=True)
    finalkeep = modelpred_ 
    return finalkeep

#######Metric Function
def Metric(IPnum,abIP,TP,FN):
    FP=abIP-TP
    TN=IPnum-abIP-FN
    R= TP/(TP+FN)
    TPR = TP/(TP+FN)
    FPR = FP/(TN+FP)
    P= TP/(TP+FP)
    return round(R, 4),round(TPR, 4),round(FPR, 4),round(P, 4)

### Batch Experimental Testing

ModelList = [dfconcat_blstm,dfconcat_blstm_cnn,dfconcat_cnn,dfconcat_cnn_blstm,dfconcat_tfxgb,dfconcat_tfrf,dfconcat_tflsvm,dfconcat_wvxgb,dfconcat_wvrf,dfconcat_wvlsvm]
# ModelList = [dfconcat_blstm_cnn]

#######Define logdf to store logs for visualization
Logdf = pd.DataFrame(columns=('ModelName','DelNum','ROC_AUC','PRC_AUC','TimeCost'))

IPnumlist = pd.concat([dfconcat_blstm["source_ip"], dfconcat_blstm["destination_ip"]], axis=0)
IPnumlist=IPnumlist.drop_duplicates()
IPnumlist.reset_index(drop=True, inplace=True)
IPnum= IPnumlist.shape[0] ##IP Num should be the same

ModelName = 0
for modelpred in ModelList:
    ModelName += 1###Model name ID, from 1 to 10, represents the above models in order

    for i in range(0,7):######Test for 0-6 ###### Of course you can increase the number of scanners to be excluded, but the results will not be better
        time_start = time.time() 

        #######Define Lists to keep TPR FPR R P
        TPRlist = []
        FPRlist = []
        Plist = []
        Rlist = []
        #######Pruning or not by changing i
        finalkeep = Pruning(modelpred,i)

        num = 1 ########Draw ROC and PRC curves according to the selection of num, That is, the number of considered abnormal records
        #When num is 0, i.e. all records are considered normal  
        #At this point R TPR FPR are 0, and P is not defined  
        #So we define num to start from 1
        while num < finalkeep.shape[0]+50:###+50 Ensure that all records can be obtained
            ab_finalsort= finalkeep[0:num]
            num += 50
            dfabip=pd.concat([ab_finalsort["source_ip"], ab_finalsort["destination_ip"]], axis=0)
            dfabip=dfabip.drop_duplicates()
            dfabip.reset_index(drop=True, inplace=True)
            if len(dfabip) == 0:
                continue
            dfabip.tolist()
            dfabip_set = set(dfabip)##Predicted set of abnormal IPs
            bingo=(dfabip_set)&(dfabcorrect_set)
            abIPi=len(dfabip_set)####Number of predicted abnormal IPs
            TPi=len(bingo)
            FNi=len(dfabcorrect)-TPi
            R,TPR,FPR,P = Metric(IPnum,abIPi,TPi,FNi)
            TPRlist.append(TPR)
            FPRlist.append(FPR)
            Plist.append(P)
            Rlist.append(R)
        ROC = pd.DataFrame(FPRlist, columns=['FPR'])
        ROC = pd.concat([ROC, pd.DataFrame(TPRlist,columns=['TPR'])],axis=1)
        PR = pd.DataFrame(Rlist, columns=['R'])
        PR = pd.concat([PR, pd.DataFrame(Plist,columns=['P'])],axis=1)

                
        ####Importance drop_duplicates
        ROC=ROC.drop_duplicates()
        PR=PR.drop_duplicates()
        ####Importance sort_values
        ROC.sort_values(by=['FPR','TPR'],inplace=True,ascending=True)
        PR.sort_values(by=['R','P'],inplace=True,ascending=[True,False])#########Same R, P starting from the higher
        ROC.reset_index(drop=True, inplace=True)
        PR.reset_index(drop=True, inplace=True)
        
        #################Add after the Sort
        #####Importance Add the last line For ROC
        ROC.loc[ROC.shape[0]] = [1,ROC.loc[ROC.shape[0]-1,'TPR']]     
        #####Importance Add the first line For PR
        PR.loc[-1] = [0,PR.loc[0,'P']]
        PR.index = PR.index + 1
        PR.sort_index(inplace=True)
        
        #####AUC
        ROC_AUC = np.trapz(list(ROC['TPR']),list(ROC['FPR'])).round(4)
        PRC_AUC = np.trapz(list(PR['P']),list(PR['R'])).round(4)
        
        time_end = time.time() 
        cost = round((time_end - time_start),4)
        
        ######Save Logs
        Path_ROC = "Results/WF_Batch_Pruning/ROC_Model%d_Prun%d.csv"%(ModelName,i)
        ROC.to_csv(Path_ROC,index=0)
        Path_PRC = "Results/WF_Batch_Pruning/PRC_Model%d_Prun%d.csv"%(ModelName,i)
        PR.to_csv(Path_PRC,index=0)
        
#         print(ModelName,i,ROC_AUC,PRC_AUC,cost)
        Logdf = Logdf.append({'ModelName':ModelName,'DelNum':i,'ROC_AUC':ROC_AUC,'PRC_AUC':PRC_AUC,'TimeCost':cost}, ignore_index=True)
#         ModelName i ROC_AUC PRC_AUC cost
########Logdf should have 10*7 records 

Logdf

Logdf.to_csv("Results/WF_Batch_Pruning/All_Logs.csv",index=0)

Logdf[Logdf['ModelName']==1]['ROC_AUC']

### Extraction for DFADR

modelpred = dfconcat_blstm_cnn
i = 3

IPnumlist = pd.concat([modelpred["source_ip"], modelpred["destination_ip"]], axis=0)
IPnumlist=IPnumlist.drop_duplicates()
IPnumlist.reset_index(drop=True, inplace=True)
IPnumi = IPnumlist.shape[0] 

time_start = time.time() 

#######Define Lists
Plist = []
Rlist = []
numlist = []
#######Pruning or not by changing i
finalkeep = Pruning(modelpred,i)

num = 1 ########Draw ROC and PRC curves according to the selection of num, That is, the number of considered abnormal records
#When num is 0, i.e. all records are considered normal  
#At this point R TPR FPR are 0, and P is not defined  
#So we define num to start from 1
while num < finalkeep.shape[0]+50:###+50 Ensure that all records can be obtained
    ab_finalsort= finalkeep[0:num]
    num += 50
    dfabip=pd.concat([ab_finalsort["source_ip"], ab_finalsort["destination_ip"]], axis=0)
    dfabip=dfabip.drop_duplicates()
    dfabip.reset_index(drop=True, inplace=True)
    if len(dfabip) == 0:
        continue
    dfabip.tolist()
    dfabip_set = set(dfabip)##Predicted set of abnormal IPs
    bingo=(dfabip_set)&(dfabcorrect_set)
    abIPi=len(dfabip_set)####Number of predicted abnormal IPs
    TPi=len(bingo)
    FNi=len(dfabcorrect)-TPi
    R,TPR,FPR,P = Metric(IPnumi,abIPi,TPi,FNi)
    
    #####Only Keep R==1
    if R==1:
        numlist.append(num-50)
        Plist.append(P)
        Rlist.append(R)

PRKeep = pd.DataFrame(Rlist, columns=['R'])
PRKeep = pd.concat([PRKeep, pd.DataFrame(Plist,columns=['P'])],axis=1)
PRKeep = pd.concat([PRKeep, pd.DataFrame(numlist,columns=['num'])],axis=1)


PRKeep.sort_values(by=['R','P'],inplace=True,ascending=True)
PRKeep.drop_duplicates(subset=['R'],keep='last', inplace=True)
PRKeep.reset_index(drop=True, inplace=True)

PRKeep

ab_final_output= finalkeep[0:PRKeep['num'][0]]
ab_final_output.to_csv("Results/WebFrontend_for_DFADR.csv",index=0)
PRKeep['num'][0]
